Minimal WindowServer and WLWindow classes
WindowServer is a wayland compositor running on drm/kms and libinput.
It is derived from tinywl from the wlroots project.

WLWindow is an implementation of CGWindow analogous to X11Window in
AppKit. It provides an abstraction of a wayland client and will eventually
replace the X11 backend in AppKit and friends.