Add getenv, setenv, unsetenv to launchctl [ci skip]
Add commands to manipulate environment variables which are needed for
dbus integration. As on macOS, variables are manipulated in the caller's
context. For regular users, entries are stored in KEY=VALUE format in
~/.launchd.env. For root, entries are stored in /etc/launchd_user.env.

Launchd will read both files when starting a new job, with per-user
values overriding system values.

This is only half the work - launchd currently only reads the system
/etc/launchd_user.env.