Move libdrm, wayland to OpenGL.framework
This commit moves libdrm and wayland to the OpenGL framework as they are dependencies
of mesa, which provides EGL and GLES to the framework. It also fixes missing files from
the libdrm build, and fixes the mesa build.

The intent here is to build a real OpenGL framework that is self-contained, rather than
linking to stuff in /usr/lib. This framework is a dependency of WindowServer.