Fix remaining Frameworks to build in OBJDIR
This commit also removes DBusKit from the build and ifdef's out the use of DBus and
XDG from LaunchServices for now. Both sections should be rewritten to use other methods.
DBus was only needed for opening files in the old Filer app and XDG is used to support
MIME database and .desktop files as apps.