Hack WindowServer into building again \o/
This commit removes the SSD, OSD, font, theme, and related
capabilities from labwc to eliminate the dependency on pango.
It also removes code that relied on glib functions which WS
does not really need.

Removing SSD definitely breaks some things, including being
able to toggle CSD/SSD for foot.app, moving and resizing SSD
windows, and the client menu. None of this is theoretically
needed for Cocoa but needs to be reimplemented for wayland
SSD clients.