Introduce LINK_SUBDIR for rvn.framework.mk
LINK_SUBDIR=subdir/path can be included in a Makefile using rvn.framework.mk
to automatically collect all .o files from the specified folder and include
them in the link of the framework's shared library. This allows easy bundling
of dependent third-party code into a framework and removes the need to link
both the framework and its dependencies when building an app against it.

Specific objects in subdir/path can be excluded by adding a line with
LINK_EXCLUDE.subdir/path= obj1.o obj2.o

The object names are relative to the subdir/path.

This commit also moves fontconfig from AppKit to CoreText.