Work around NSSelectInputSource oddity
Something weird is happening with -[NSSelectInputSource socket].

Although init takes a NSSocket* and socket returns a NSSocket*, a call to
[NSSelectInputSource socket] is somehow returning the value as an
NSSocketNativeHandle, which is an int. This means the 64-bit addr of the
socket is missing its upper half and causes a crash.

NSSocketNativeHandle is only used in NSSocketPort's socket method so it
appears that somehow the runtime is confusing the class of the object?
Something to look at later.