bridge: be consistent about PVID terminology
The PVID option is exposed to the user as 'untagged', but the API was
inconsistent on whether it's called 'untagged' or 'PVID'.  Standardise
on calling it PVID everywhere in the code, since this is the 802.1Q
terminology.

Keep 'untagged' as the user-facing term since sysadmins are not network
admins and are often not familiar with the term PVID.

Approved by:	des
Differential Revision:	https://reviews.freebsd.org/D51182

(cherry picked from commit 4d281507f5f159e0717ae9f4a46cf6fbca67e23d)