LinuxKPI: pci: fix argument type to linuxkpi_pci_iomap[_range]
The last argument (maxlen) to linuxkpi_pci_iomap_range and
linuxkpi_pci_iomap is an unsigned long not an int.  LinuxKPI is
not using that argument in the end but fix it where needed.
While here adjust the name to 'maxlen' and remove the "mmio_" to
bar and off.

Sponsored by:	The FreeBSD Foundation
MFC after:	3 days
Reviewed by:	emaste
Differential Revision: https://reviews.freebsd.org/D51651

(cherry picked from commit 421bf3a968674cbb1fc7eb1c323fd4d5d2409718)